Move calendar to own analytics folder
